Fix state visibility, styleguide and variable type consistency

+++ project-examples/ethereum/contracts/Design Patterns/Pull Pattern/PullPattern.sol
@@ -3,9 +3,9 @@
 pragma solidity 0.7.0;
 
 contract PullPattern {
-    mapping (address => uint) public balances;
-    event AccountBalance(uint balance);
-    
+    mapping(address => uint256) public balances;
+    event AccountBalance(uint256 balance);
+
     receive() external payable {
         balances[msg.sender] = msg.value;
     }
@@ -17,7 +17,7 @@ contract PullPattern {
     function payout() public {
         require(balances[msg.sender] > 0, "No balance available.");
 
-        uint amount = balances[msg.sender];
+        uint256 amount = balances[msg.sender];
         balances[msg.sender] = 0;
 
         (bool success, ) = msg.sender.call{value: amount}("");
